requirements.txt
Events~=0.5
paho-mqtt~=2.1.0
PyYAML~=6.0.2
